Included a new form
There is now a new class, which represents a form window, that allows
you to adjust the orientation of a point cloud. This is to let a user
calibrate and position a point cloud, accordingly, against another one.

Additionally, there is also a new class, MultiFormContext, which manages
the multiple, simultaneously running forms.

Lastly, there is a new method in PointCloudManipulation that allows
input from the new form to facilitate adjusting.